backend/requirements.txt
# Core web framework
fastapi==0.116.2
uvicorn==0.35.0
pydantic==2.11.9
python-dotenv==1.1.1
python-multipart==0.0.9

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

# Authentication
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0

# Google AI (required for AI service)
google-generativeai>=0.8.0

# gRPC for microservices (runtime only, not tools)
grpcio>=1.62.0,<1.76.0
grpcio-status>=1.62.0,<1.76.0
protobuf>=4.21.0,<6.0.0

# HTTP client for embedding service
requests>=2.32.0

# Email
resend==2.4.0

# Logging
colorlog==6.8.2
